Fix warning-as-error diagnostics in build and specs
Agent-Logs-Url: https://github.com/Cratis/Chronicle/sessions/31ebf916-6071-4f44-9bc2-1fc7faef5226

+++ Source/Kernel/Storage.Sql/EventStores/Namespaces/UniqueConstraints/UniqueConstraintsStorage.cs
@@ -44,10 +44,10 @@ public class UniqueConstraintsStorage(EventStoreName eventStore, EventStoreNames
         {
             if (existing.EventSourceId == eventSourceId)
             {
-                return (true, (EventSequenceNumber)existing.SequenceNumber);
+                return (true, (EventSequenceNumber)(ulong)existing.SequenceNumber);
             }
 
-            return (false, (EventSequenceNumber)existing.SequenceNumber);
+            return (false, (EventSequenceNumber)(ulong)existing.SequenceNumber);
         }
 
         return (true, EventSequenceNumber.Unavailable);
